requirements.txt
google-cloud-storage 
langchain
flask-cors 